requirements.txt
gradio[mcp]
anthropic>=0.7.0
mistralai
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10
aiofiles>=23.0.0
pydantic>=2.0.0
httpx>=0.24.0
uvicorn>=0.23.0
python-multipart>=0.0.6
asyncio-mqtt>=0.11.1
gradio[mcp]
anthropic>=0.7.0
mistralai
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10
aiofiles>=23.0.0
pydantic>=2.0.0
httpx>=0.24.0
uvicorn>=0.23.0
python-multipart>=0.0.6
asyncio-mqtt>=0.11.1
nest-asyncio>=1.5.6
httpx
fastmcp
mcp
openai
python-dotenv
llama-index
llama-index-llms-openai
llama-index-llms-anthropic
llama-index-embeddings-huggingface

# ElevenLabs Voice Integration
elevenlabs>=1.0.0
websockets>=12.0